cec-uinput: don't blindly activate display
The bActivateSource flag says to wake up the television, and this
has bad consequences when we re-boot machines overnight.

This should be done only if the power to the television is on,
and running cec-uinput interactively shows this when a TV is
powered off:

	power status changed from 'unknown' to 'standby'

or this when powered on:
	power status changed from 'unknown' to 'on'

So a follow-up patch could activate the display if and only if
the power is on.

Signed-off-by: Eric Nelson <eric@nelint.com>