cec-uinput: activate device on power-up
During a power cycle, my Samsung television will switch to
the non-CEC enabled input and because of this, it won't receive
keypresses until I manually select the "MPlayer" (CEC-enabled)
input source using the television's remote control.

To prevent this, select the MPlayer input whenever we receive
a "power-on" message from the television.

Signed-off-by: Eric Nelson <eric@nelint.com>