cec-uinput: only activate after power-on
Commits d881b65 and b85a5c5 were a bit over-zealous and would
respond to routing messages regardless of whether the TV was
previously powered on, which resulted in an automatic switch
to our input when switching to non-CEC channels.

Keep track of the state of the TV power (unknown, off, or on)
and only actively select our input when transitioning from
unknown or off.

Signed-off-by: Eric Nelson <eric@nelint.com>